+++ changelog.md
@@ -232,3 +232,4 @@
 * Added modded support to `ChoppingTask` as well as several optimizations
 * Mining Speed Multipliers can now effect `ChoppingTask` speed (The original 7 seconds is also configurable)
 * Fixed potential crashes when villagers perform Harvesting chores (Planting seeds throwing a NPE)
+* Fixed player marriage not saving